Fix format string warning in exit logging
FIXES:
- Change %ld to %lld for time_t timestamp in exit logging
- Cast time(NULL) to (long long) to match format specifier
- Resolves compilation warning about format string mismatch
- Ensures proper timestamp display in exit messages

The exit logging system should now compile without warnings.

+++ src/c/autonomy-daemon/core/autonomy-daemon.c
@@ -348,7 +348,7 @@ static void log_exit_reason(exit_reason_t reason, const char *message) {
     fprintf(stderr, "\n=== DAEMON EXIT ===\n");
     fprintf(stderr, "Exit Reason: %s\n", reason_str);
     fprintf(stderr, "Exit Message: %s\n", g_exit_message);
-    fprintf(stderr, "Timestamp: %ld\n", time(NULL));
+    fprintf(stderr, "Timestamp: %lld\n", (long long)time(NULL));
     fprintf(stderr, "PID: %d\n", getpid());
     fprintf(stderr, "==================\n\n");
